Adds skills section with interactive tooltips
Introduces a new section on the homepage to display a grid of professional skills and tools, enhancing the portfolio. Each skill icon now features an interactive tooltip on hover to display its name.

This change integrates Radix UI for the tooltip component and adds necessary utility packages like `cva`, `clsx`, and `tailwind-merge`.

Additionally, the main layout is updated to be wider with adjusted responsive breakpoints for better viewing on larger screens. The project structure is also improved by refactoring UI components and removing unused assets.